feat: change structure, new layout
