+++ src/gamefield/players/buttons/buttonsSectionRight.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function ButtonSectionRight() {
+    return(
+        <div className="buttons-section-right">
+            <button className="button save-game-button">Сохранить игру</button>
+            <button className="button new-game-button">Новая игра</button>
+            <button className="button end-game-button">Завершить игру</button>
+        </div>
+    )
+}
+
+export default ButtonSectionRight;
