+++ src/index.css
@@ -28,6 +28,8 @@ code {
 
 .GAME_FIELD_WRAPPER{
   display: flex;
+  flex-direction: column;
+  justify-content: space-around;
   height: 95%;
 }
 
@@ -48,17 +50,27 @@ code {
 }
 
 .host-section {
+  position: relative;
   display: flex;
-  flex-direction: column;
   justify-content: center;
-  align-items: center;
-  width: 30%;
-  height: 100%
+  align-items: flex-end;
+  width: 50%;
+  height: 95%
 }
 
 .game-section {
-  width: 70%;
-  height: 100%;
+  margin: 0 auto;
+  width: 95%;
+  height: 63%;
+}
+
+.bottom-section {
+  margin: 0 auto;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 95%;
+  height: 30%;
 }
 
 .questions-section {
@@ -66,8 +78,8 @@ code {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  width: 70%;
-  height: 60%;
+  width: 100%;
+  height: 90%;
 }
 
 .round-number {
@@ -94,14 +106,14 @@ code {
   display: flex;
   justify-content: flex-start;
   align-items: center;
-  width: 37.5%;
+  width: 50%;
 }
 
 .question {
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 12.5%;
+  width: 10%;
   border-left: 5px solid yellow;
 }
 
@@ -113,9 +125,9 @@ code {
 .players-section {
   display: flex;
   justify-content: space-around;
-  align-items: center;
-  width: 70vw;
-  height: 40vh;
+  align-items: flex-end;
+  width: 50%;
+  height: 95%;
 }
 
 .player-icon {
@@ -131,9 +143,9 @@ code {
 }
 
 .info {
-  position: fixed;
-  top: 20%;
-  left: 17%;
+  position: absolute;
+  top: -20%;
+  left: 60%;
   width: 350px;
   height: 250px;
   background-image: url("./img/Speech_bubble.svg.png");
@@ -152,11 +164,20 @@ code {
 
 .buttons-section {
   display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 360px;
+  height: 250px;
+}
+
+.buttons-section-left,
+.buttons-section-right {
+  display: flex;
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
   width: 170px;
-  height: 170px;
+  height: 250px;
 }
 
 .button {
